feat!: Replace Alias with model on propertyValue
BREAKING:

Removes the duplicate Alias property on the PropertyValue model. This can be retrived from the container model and is therefore obsolete. The new Model property is the replacement showing the model that the property value is using. This was added because the interfaceType requires at least one public property.